Change type of volume property from int to qreal for SoundEffect Element
Task-Number: QTMOBILITY-590
(cherry picked from commit c177b9c110246d7e59d02c65e9471c5260b9cdaa)

Conflicts:

	src/multimedia/effects/qsoundeffect.cpp

+++ src/multimedia/effects/qsoundeffect_p.h
@@ -71,7 +71,7 @@ class Q_MULTIMEDIA_EXPORT QSoundEffect : public QObject
     Q_CLASSINFO("DefaultMethod", "play()")
     Q_PROPERTY(QUrl source READ source WRITE setSource NOTIFY sourceChanged)
     Q_PROPERTY(int loops READ loopCount WRITE setLoopCount NOTIFY loopCountChanged)
-    Q_PROPERTY(int volume READ volume WRITE setVolume NOTIFY volumeChanged)
+    Q_PROPERTY(qreal volume READ volume WRITE setVolume NOTIFY volumeChanged)
     Q_PROPERTY(bool muted READ isMuted WRITE setMuted NOTIFY mutedChanged)
     Q_PROPERTY(bool playing READ isPlaying NOTIFY playingChanged)
     Q_PROPERTY(Status status READ status NOTIFY statusChanged)
@@ -103,8 +103,8 @@ public:
     int loopCount() const;
     void setLoopCount(int loopCount);
 
-    int volume() const;
-    void setVolume(int volume);
+    qreal volume() const;
+    void setVolume(qreal volume);
 
     bool isMuted() const;
     void setMuted(bool muted);
